Simple Cipher
Started by uploading my arcade model designed in Fusion 360 onto Unity and started decorating it with material colors. However, I found out that there is a problem with the main knob where it is difficult to see the base letters behind the knob. I am going to have to redesign this to be able to have a better visual.